Remove EDH fitness totals api methods & tests
