--- source/api/fitness-totals/consts/fitness-types.js
@@ -1,1 +1,0 @@
-export default ['bike', 'climb', 'gym', 'hike', 'run', 'sport', 'swim', 'walk']
